Improved speed and fixed overwrite of forces
When using the independent method, the apply_forces() method in
restraint was zeroeing the forces, so that only the last applied
forces actually worked. Moving the zeroing out of the main loop also
improved the speed. Also, I didn't realize apply_forces() used the
number of atoms in each CV, so I had to move the swaping of caches
atom numbers down past that.